+++ src/SWCForecastBase.jl
@@ -1,10 +1,11 @@
 module SWCForecastBase
 
-# Write your package code here.
+# # Imputation
 using Impute, Dates, Statistics, DataFrames
 include("myimputation/myimpute.jl")
 export imputemean!, imputeinterp!, removeunreasonables!
 
+# # Fundamental utilities
 
 include("myimputation/checkmissnan.jl")
 export chknnm, isnnm, islnan
@@ -16,27 +17,42 @@ using Chain
 include("combinegroup.jl")
 export combinegroup_allcols
 
+using NaNMath
+include("precipitation.jl")
+export movingaverage
+
+# # Series to supervised
+
 using ShiftedArrays
 include("series2supervised.jl")
 export series2supervised
 
-
-
-using NaNMath
-include("precipitation.jl")
-export movingaverage
+# # MLJ Model wrappers
 
 using MLJ
 include("mljmodels/treemodels.jl")
 export fstree, twofstree, manytrees
 
-include("traintest.jl")
+# # Prepare the table
 
 using DataFrames
-include("preparetable.jl")
+include("preparetable/briefinfo.jl")
+include("preparetable/preparetable0.jl")
+include("preparetable/preparetable.jl")
+
 export PrepareTable, PrepareTableDefault, preparetable!, ConfigAccumulate, ConfigPreprocess, ConfigSeriesToSupervised
 
+# # Train and test
 
+using OkTableTools
+include("preparetable/traintest.jl")
+export traintest!, train!, test!
+
+using CSV, Random
+include("preparetable/save.jl")
+export save
+
+# # Plotting
 
 include("forplot/dataoverview.jl") # only for test